requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
mysqlclient>=2.2,<3.0
django-cors-headers>=4.4,<5.0
Pillow>=10.0,<11.0

